Share the simulators' payment and disallow-list helpers from helix-common
The disallow list's parsing and digest, and the Safe multiSend payment
recognition, carry no reth types. Move them so the ethrex simulator uses
the same code. The functions and their tests move verbatim.

Step 2 of gattaca-com/helix#527.
