+++ crates/common/Cargo.toml
@@ -10,6 +10,7 @@ version.workspace = true
 alloy-consensus.workspace = true
 alloy-primitives.workspace = true
 alloy-rlp.workspace = true
+alloy-sol-types.workspace = true
 axum.workspace = true
 backtrace.workspace = true
 bytes.workspace = true
