+++ crates/common/src/lib.rs
@@ -6,6 +6,7 @@ pub mod api;
 pub mod api_provider;
 pub mod beacon;
 pub mod bid_submission;
+pub mod blacklist;
 pub mod builder_info;
 pub mod chain_info;
 pub mod config;
@@ -13,6 +14,7 @@ pub mod decoder;
 pub mod http;
 pub mod local_cache;
 pub mod metrics;
+pub mod payment;
 pub mod proposer;
 pub mod signing;
 pub mod simulator;
