+++ crates/simulator/Cargo.toml
@@ -40,7 +40,6 @@ revm.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 serde_with.workspace = true
-sha2.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 tracing.workspace = true
